+++ src/databricks/labs/remorph/config.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 
 from sqlglot.dialects.dialect import Dialect, Dialects, DialectType
 
-from databricks.labs.remorph.helpers.morph_status import ParserError
+from databricks.labs.remorph.transpiler.transpile_status import ParserError
 from databricks.labs.remorph.reconcile.recon_config import Table
 from databricks.labs.remorph.transpiler.sqlglot.generator import databricks
 from databricks.labs.remorph.transpiler.sqlglot.dialects.snowflake import snowflake
