+++ src/databricks/labs/remorph/transpiler/sqlglot/sqlglot_engine.py
@@ -5,7 +5,7 @@ from sqlglot.expressions import Expression
 
 from databricks.labs.remorph.config import TranspilationResult
 from databricks.labs.remorph.helpers.file_utils import refactor_hexadecimal_chars
-from databricks.labs.remorph.helpers.morph_status import ParserError
+from databricks.labs.remorph.transpiler.transpile_status import ParserError
 
 
 class SqlglotEngine:
